+++ packages/react-core/src/filters/FilterQueryBuilder.d.ts
@@ -1,6 +1,0 @@
-export enum FilterTypes {
-  IN = 'in',
-  BETWEEN = 'between', // [a, b] both are included
-  CLOSED_OPEN = 'closed_open', // [a, b) a is included, b is not
-  TIME = 'time'
-}