+++ packages/react-core/src/filters/FilterTypes.d.ts
@@ -1,0 +1,16 @@
+export enum FilterTypes {
+  IN = 'in',
+  BETWEEN = 'between', // [a, b] both are included
+  CLOSED_OPEN = 'closed_open', // [a, b) a is included, b is not
+  TIME = 'time',
+  STRING_SEARCH = 'stringSearch'
+}
+
+export const filterFunctions: Record<
+  string,
+  (
+    filterValues: unknown[],
+    featureValue: unknown,
+    params?: Record<string, unknown>
+  ) => boolean
+>;